pyproject.toml
[build-system]
requires = ["setuptools >= 77.0.3"]
build-backend = "setuptools.build_meta"

[project]
name = "f1dbtduckdb"
version = "0.6.0"
description = "Experimenting with dbt/DuckDB using Ergast API as main source"
readme = "README.md"
requires-python = ">= 3.10"
urls.homepage = "https://github.com/frizzleqq/f1-dbt-duckdb"
classifiers = [
  "Natural Language :: English",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]

dependencies = [
  "dagster-duckdb>=0.23.0",
  "dagster-dbt>=0.23.0",
  "dagster-webserver>=1.7.0,<2.0.0",
  "dbt-duckdb>=1.8.0,<2.0.0",
  "duckdb>=1.0.0,<2.0.0",
]

[project.optional-dependencies]
dev = [
  # development & testing tools
  "mypy>=1.0.0, <2.0.0",
  "ruff>=0.5.0",
  "sqlfluff-templater-dbt>=3.0.0, <4.0.0",
  "build>=1.0.0, <2.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["foneplatform*", "dbt*"]

[tool.setuptools.package-data]
dbt = [
  "*.yml",
  "macros/*",
  "models/core/*",
  "models/export/*",
  "models/stage/ergast/*",
]

[tool.dagster]
module_name = "foneplatform"
code_location_name = "foneplatform"

[[tool.mypy.overrides]]
module = "ergast,staging,requests.adapters"
ignore_missing_imports = true

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = [
  "E", # pycodestyle
  "W", # pycodestyle
  "F", # Pyflakes
  "I", # isort
  "T201", # print statements
]
ignore = [
  "E501", # Line length is regulated by formatter
]

[tool.ruff.lint.pydocstyle]
convention = "numpy"
